+++ src/client/processor/online_push.rs
@@ -4,7 +4,7 @@ use crate::client::income::decoder::online_push::GroupMessagePart;
 impl Client {
     pub async fn process_group_message_part(&self, group_message_part: GroupMessagePart) {
         println!("{:?}", group_message_part);
-        self.mark_group_message_readed(group_message_part.group_code, group_message_part.seq as i64).await;
+        self.mark_group_message_readed(group_message_part.group_code, group_message_part.seq).await;
         // TODO merge part and dispatch to handler
     }
 }
